+++ lib/ui/page/home/page/chat/controller.dart
@@ -147,8 +147,7 @@ class ChatController extends GetxController {
   /// typing in this [chat].
   StreamSubscription? _typingSubscription;
 
-  /// List of [StreamSubscription] that indicates interest of this controller
-  ///  in [User]s updates
+  /// [StreamSubscription]s to the [RxUser.updates] of the [RxChat.members].
   List<StreamSubscription>? _usersSubscriptions;
 
   /// Indicator whether [_updateFabStates] should not be react on
@@ -267,27 +266,21 @@ class ChatController extends GetxController {
         }
       },
     );
+
     super.onInit();
   }
 
   @override
   void onReady() {
     listController.addListener(_updateFabStates);
-    _fetchChat().then((_) {
-      _usersSubscriptions = chat?.members.values
-          .where((u) => u.user.value.id != me)
-          .map((rxUser) => rxUser.updates.listen((event) {}))
-          .toList();
-    });
+    _fetchChat();
     _initAudio();
     super.onReady();
   }
 
   @override
   void onClose() {
-    _usersSubscriptions?.forEach((subscription) {
-      subscription.cancel();
-    });
+    _usersSubscriptions?.forEach((s) => s.cancel());
     _messagesWorker?.dispose();
     _readWorker?.dispose();
     _typingSubscription?.cancel();
@@ -492,6 +485,11 @@ class ChatController extends GetxController {
       if (lastRead?.value.id != lastReadItem.value?.value.id) {
         _scrollToLast();
       }
+
+      _usersSubscriptions = chat?.members.values
+          .where((u) => u.user.value.id != me)
+          .map((u) => u.updates.listen((_) {}))
+          .toList();
 
       status.value = RxStatus.success();
     }