+++ lib/ui/page/home/widget/user_search_bar/controller.dart
@@ -89,10 +89,10 @@ class UserSearchBarController extends GetxController {
   }
 
   /// Adds the provided [user] to the [recentSearchResults].
-  void addToRecent(RxUser rxUser) {
+  void addToRecent(RxUser user) {
     recentSearchResults
-        .removeWhere((e) => e.user.value.id.val == rxUser.user.value.id.val);
-    recentSearchResults.add(rxUser);
+        .removeWhere((e) => e.user.value.id.val == user.user.value.id.val);
+    recentSearchResults.add(user);
     if (recentSearchResults.length >= 10) {
       recentSearchResults.removeAt(0);
     }